test: remove e2e optimization changes
