+++ tests/suites/tenant/diagnostics/Diagnostics.ts
@@ -11,12 +11,6 @@ import {OperationsTable} from './tabs/OperationsModel';
 
 const OWNER_CARD_VISIBILITY_TIMEOUT = VISIBILITY_TIMEOUT * 2;
 
-async function getRowCellsText(row: Locator): Promise<string[]> {
-    const cells = row.locator('td');
-    const cellTexts = await cells.allTextContents();
-    return cellTexts.map((text) => text.replace(/\s+/g, ' ').trim());
-}
-
 export enum DiagnosticsTab {
     Info = 'overview',
     Database = 'database',
@@ -403,7 +397,8 @@ export class Diagnostics {
 
     async getRowData(rowIndex: number): Promise<string[]> {
         const row = this.dataTable.locator(`tbody tr:nth-child(${rowIndex + 1})`);
-        return getRowCellsText(row);
+        const cells = row.locator('td');
+        return await cells.allTextContents();
     }
 
     async clickRefreshButton(): Promise<void> {
@@ -530,23 +525,11 @@ export class Diagnostics {
         return rowElementClass?.includes('kv-top-queries__row_active') || false;
     }
 
-    async waitForActiveRowData(expectedRowData: string[]): Promise<void> {
-        const activeRow = this.dataTable
+    async waitForActiveRow(): Promise<void> {
+        await this.dataTable
             .locator('tr.data-table__row.kv-top-queries__row_active')
-            .first();
-
-        await retryAction(async () => {
-            await activeRow.waitFor({state: 'visible', timeout: VISIBILITY_TIMEOUT});
-            const activeRowData = await getRowCellsText(activeRow);
-
-            if (JSON.stringify(activeRowData) !== JSON.stringify(expectedRowData)) {
-                throw new Error(
-                    `Active row data ${JSON.stringify(
-                        activeRowData,
-                    )} did not match expected ${JSON.stringify(expectedRowData)}`,
-                );
-            }
-        });
+            .first()
+            .waitFor({state: 'visible', timeout: VISIBILITY_TIMEOUT});
     }
 
     async isOwnerCardVisible(timeout = OWNER_CARD_VISIBILITY_TIMEOUT): Promise<boolean> {
