+++ tests/suites/tenant/diagnostics/tabs/queries.test.ts
@@ -271,7 +271,6 @@ test.describe('Diagnostics Queries tab', async () => {
 
         // Target a row further down that requires scrolling
         const targetRowIndex = 8;
-        const targetRowData = await diagnostics.getRowData(targetRowIndex - 1);
 
         // Click on the target row to open the drawer
         await diagnostics.table.clickRow(targetRowIndex);
@@ -293,6 +292,6 @@ test.describe('Diagnostics Queries tab', async () => {
         await page.waitForTimeout(1000);
 
         // Verify the selected row is highlighted and visible.
-        await diagnostics.waitForActiveRowData(targetRowData);
+        await diagnostics.waitForActiveRow();
     });
 });
